miden_prover::crypto

Module merkle

Structs

EmptySubtreeRoots

Contains precomputed roots of empty subtrees in a Merkle tree.

InnerNodeInfo

Representation of a node with two children used for iterating over containers.

LeafIndex

The index of a leaf, at a depth known at compile-time.

MerklePath

A merkle path container, composed of a sequence of nodes of a Merkle tree.

MerkleStore

An in-memory data store for Merkelized data.

MerkleTree

A fully-balanced binary Merkle tree (i.e., a tree where the number of leaves is a power of two).

Mmr

A fully materialized Merkle Mountain Range, with every tree in the forest and all their elements.

MmrPeaks

NodeIndex

Address to an arbitrary node in a binary tree using level order form.

PartialMerkleTree

A partial Merkle tree with NodeIndex keys and 4-element Word leaf values. Partial Merkle Tree allows to create Merkle Tree by providing Merkle paths of different lengths.

SimpleSmt

A sparse Merkle tree with 64-bit keys and 4-element leaf values, without compaction.

Smt

Sparse Merkle tree mapping 256-bit keys to 256-bit values. Both keys and values are represented by 4 field elements.

SmtProof

A proof which can be used to assert presence (or absence) of key-value pairs in a super::Smt (Sparse Merkle Tree).

StoreNode

Enums

MerkleError

SmtProofError

Errors that can occur when validating SMT proofs.

Constants

SMT_DEPTH

The depth of the sparse Merkle tree.