miden_prover::crypto::random

Struct RandomCoin

pub struct RandomCoin { /* private fields */ }

A simplified version of the SPONGE_PRG reseedable pseudo-random number generator algorithm described in https://eprint.iacr.org/2011/499.pdf.

The simplification is related to the following facts:

A call to the reseed method implies one and only one call to the permutation function. This is possible because in our case we never reseed with more than 4 field elements.

As a result of the previous point, we don’t make use of an input buffer to accumulate seed material.

Implementations

impl RandomCoin

pub fn new(seed: Word) -> RandomCoin

Returns a new RandomCoin initialized with the specified seed.

pub fn from_parts(state: [Felt; 12], current: usize) -> RandomCoin

Returns a RandomCoin instantiated from the provided components.

Panics

Panics if current is outside of the rate range.

pub fn into_parts(self) -> ([Felt; 12], usize)

Returns components of this random coin.

pub fn fill_bytes(&mut self, dest: &mut [u8])

Fills dest with random data.

pub fn draw_basefield(&mut self) -> Felt

Draws a random base field element from the random coin.

This method applies the Poseidon2 permutation when the rate portion of the state is exhausted, then returns the next element from the rate portion.

pub fn draw(&mut self) -> Felt

Draws a random field element.

This is an alias for Self::draw_basefield.

pub fn draw_ext_field<E>(&mut self) -> E
where E: ExtensionField<Felt>,

Draws a random extension field element.

The extension field element is constructed by drawing E::DIMENSION base field elements and interpreting them as basis coefficients.

pub fn reseed(&mut self, data: Word)

Reseeds the random coin with additional entropy.

The provided data is added to the first half of the rate portion of the state, then the Poseidon2 permutation is applied. The buffer pointer is reset to the start of the rate portion.