miden_prover::field

Struct BinomialExtensionField

pub struct BinomialExtensionField<F, const D: usize, A = F> { /* private fields */ }

Implementations

impl<F, A, const D: usize> BinomialExtensionField<F, D, A>

pub const fn new(value: [A; D]) -> BinomialExtensionField<F, D, A>

Create an extension field element from an array of base elements.

Any array is accepted. No reduction is required since base elements are already valid field elements.

impl<F, const D: usize> BinomialExtensionField<F, D>
where F: Copy,

pub const fn new_array<const N: usize>( input: [[F; D]; N], ) -> [BinomialExtensionField<F, D>; N]

Convert a [[F; D]; N] array to an array of extension field elements.

Const version of input.map(BinomialExtensionField::new).

Panics

Panics if N == 0.

impl<R> BinomialExtensionField<R, 2>
where R: PrimeCharacteristicRing,

Convenience methods for complex extensions

pub const fn new_complex(real: R, imag: R) -> BinomialExtensionField<R, 2>

pub const fn new_real(real: R) -> BinomialExtensionField<R, 2>

pub const fn new_imag(imag: R) -> BinomialExtensionField<R, 2>

pub fn real(&self) -> R

pub fn imag(&self) -> R

pub fn conjugate(&self) -> BinomialExtensionField<R, 2>

pub fn norm(&self) -> R

pub fn to_array(&self) -> [R; 2]

pub fn rotate<Ext>( &self, rhs: &BinomialExtensionField<Ext, 2>, ) -> BinomialExtensionField<Ext, 2>
where Ext: Algebra<R>,