miden_prover

Function prove

pub async fn prove(
    program: &Program,
    stack_inputs: StackInputs,
    advice_inputs: AdviceInputs,
    host: &mut impl Host,
    options: ProvingOptions,
) -> Result<(StackOutputs, ExecutionProof), ExecutionError>

Executes and proves the specified program and returns the result together with a STARK-based proof of the program’s execution.

This is an async function that works on all platforms including wasm32.

stack_inputs specifies the initial state of the stack for the VM.

host specifies the host environment which contain non-deterministic (secret) inputs for the prover

options defines parameters for STARK proof generation.

Errors

Returns an error if program execution or STARK proof generation fails for any reason.