miden_prover

Trait FutureMaybeSend

pub trait FutureMaybeSend<O>: Future<Output = O> + Send { }

Alias for a Future

Unless the compilation target family is wasm, we add Send to the required bounds. For wasm compilation targets there is no Send bound.